+++ database/mockdatabase.go
@@ -2,6 +2,7 @@ package database
 
 import (
 	"github.com/DarthHater/bored-board-service/model"
+	"github.com/DarthHater/bored-board-service/constants"
 )
 
 type MockDatabase struct {
@@ -27,7 +28,7 @@ func (m *MockDatabase) GetPosts(threadId string) ([]model.Post, error) {
 }
 
 func (m *MockDatabase) GetUser(username string) (result model.User, err error) {
-	result = model.User{ID: "1", Username: "CoolGuy420", EmailAddress: "hsimpson@springfield.org", UserPassword: []byte("fake password"), UserRole: int(Admin)}
+	result = model.User{ID: "1", Username: "CoolGuy420", EmailAddress: "hsimpson@springfield.org", UserPassword: []byte("fake password"), UserRole: int(constants.Admin)}
 
 	return result, nil
 }
